+++ src/test/java/SMU/BAMBOO/Hompage/domain/libraryPost/service/LibraryPostServiceImplTest.java
@@ -8,8 +8,8 @@ import SMU.BAMBOO.Hompage.domain.member.entity.Member;
 import SMU.BAMBOO.Hompage.domain.tag.entity.Tag;
 import SMU.BAMBOO.Hompage.global.exception.CustomException;
 import SMU.BAMBOO.Hompage.global.exception.ErrorCode;
-import SMU.BAMBOO.Hompage.mock.FakeLibraryPostRepository;
-import SMU.BAMBOO.Hompage.mock.FakeTagRepository;
+import SMU.BAMBOO.Hompage.mock.repository.FakeLibraryPostRepository;
+import SMU.BAMBOO.Hompage.mock.repository.FakeTagRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
